requirements.txt
gunicorn
flask
flask-cors
google-generativeai
python-dotenv
numpy
